Create base page list class
